Platforms: remove windows, clean up darwin
