+++ bot/log.go
@@ -1,5 +1,3 @@
-// +build darwin dragonfly freebsd linux netbsd openbsd
-
 package bot
 
 import (
